+++ guillotina_elasticsearch/tests/fixtures.py
@@ -27,8 +27,15 @@ def base_settings_configurator(settings):
             "sniffer_timeout": None
         }
     }
+
+    if os.environ.get('ES_VERSION', '6') == 'OPEN_DISTRO':
+        settings['elasticsearch']["connection_settings"].update({
+            "use_ssl": True,
+            "http_auth": "admin:admin"
+        })
+
     settings["load_utilities"]["catalog"] = {
-        "provides": "guillotina_elasticsearch.utility.IElasticSearchUtility",  # noqa
+        "provides": "guillotina_elasticsearch.interfaces.IElasticSearchUtility",  # noqa
         "factory": "guillotina_elasticsearch.utility.ElasticSearchUtility",
         "settings": {}
     }