+++ CHANGELOG.rst
@@ -4,6 +4,9 @@
 - Not loading utility by default and avoid crash when utility is not configured
   [bloodbare]
 
+- Support for opendistro
+  [vangheem]
+
 
 3.2.6 (2019-02-18)
 ------------------